Export mechanism for errs and associated times
If saveResultsArm is 1 in main.py, two data files are saved when the arm
closes; one with error values and one with associated times that run was
called for the arm.
Yet to be tested, the files have unique suffixes and will be overwritten
during a normal run or at the beginning of a bash sequence.
They will be appended if main.py is called via the runsim script.
Theoretically...